Move hanger/mount to the actual back!
Bit of a whoopsy in the previous design, putting the mounting holes
attached to the front. Putting them on the back plate is not ideal,
since that is only clipping in to the front case portion, so there's a
risk the main bit falls off the wall leaving only the back attached...

Revised and increased the size of the clips at the top to be entire
notches to hopefully take more weight, and added a third clip down each
of the remaining sides.

Time will tell...